backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
reportlab
pandas
kaggle
